Work in progress
Separates metadata test
